% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_list_extract.R
\name{DEG_list_extract}
\alias{DEG_list_extract}
\title{DEG_list_extract}
\usage{
DEG_list_extract(
  DEGs_list,
  log_fold_change_cut_off = 2,
  q_value_cut_off = 0.05
)
}
\arguments{
\item{DEGs_list}{needs analyzed list for extraction of DEGs}

\item{log_fold_change_cut_off}{logFC cut off value. the default value is absolute value 2}

\item{q_value_cut_off}{FDR cut off value,  the default value is absolute value 0.05}
}
\value{
a list containing 2 lists
}
\description{
To extract DEG list for further GO, KEGG enrichment analysis and vocanol plot painting
}
\details{
To extract DEG list for further GO, KEGG enrichment analysis and vocanol plot painting
}
