% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_draw.R
\name{GO_draw}
\alias{GO_draw}
\title{GO_draw}
\usage{
GO_draw(
  GO_list,
  term_num = 15,
  gene_num = 3,
  axis_text_size = 12,
  axis_title_size = 14
)
}
\arguments{
\item{GO_list}{needs GO enrichment analysis result}

\item{term_num}{needs a cut off number. To draw a tidy graph, only top 15 terms in MF, CC and BP, respectively, will be selected to draw the graph. 15 is default number which users can set the num.}

\item{gene_num}{minimum number of gene in one GO term, the default value is 3}

\item{axis_text_size}{axis_text_size can adjust the size of words in axis}

\item{axis_title_size}{axis_title_size can adjust the size of words in title}
}
\value{
a list of 2 graphs, GO_up and GO_down
}
\description{
Draw GO enrichment analysis results
}
\details{
Draw GO enrichment analysis results
}
