% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_enrich.R
\name{GO_enrich}
\alias{GO_enrich}
\title{GO_enrich}
\usage{
GO_enrich(DEG_results)
}
\arguments{
\item{DEG_results}{needs analyzed list for extraction of DEGs}
}
\value{
a dataframe
}
\description{
Perform GO enrichment analysis
}
\details{
GO enrichment analysis using fisher's exact test and multi test, respectively
}
