% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_draw.R
\name{KEGG_draw}
\alias{KEGG_draw}
\title{KEGG_draw}
\usage{
KEGG_draw(KEGG_list, axis_text_size = 14, axis_title_size = 16)
}
\arguments{
\item{KEGG_list}{needs KEGG enrichment analysis result}

\item{axis_text_size}{axis_text_size can adjust the size of words in axis}

\item{axis_title_size}{axis_title_size can adjust the size of words in title}
}
\value{
a list of 2 graphs, GO_up and GO_down
}
\description{
Draw KEGG enrichment analysis results
}
\details{
Draw KEGG enrichment analysis results
}
