% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_enrich.R
\name{KEGG_enrich}
\alias{KEGG_enrich}
\title{KEGG_enrich}
\usage{
KEGG_enrich(DEG_results)
}
\arguments{
\item{DEG_results}{needs analyzed list for extraction of DEGs}
}
\value{
a dataframe
}
\description{
Perform KEGG enrichment analysis
}
\details{
KEGG enrichment analysis using fisher's exact test and multi test, respectively
}
