% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msu_trans.R
\name{msu_trans}
\alias{msu_trans}
\title{msu_trans}
\usage{
msu_trans(DEGs_list)
}
\arguments{
\item{DEGs_list}{needs a method to DEGs_list and the rownames should be RAP-ID for MSU ID conversion}
}
\value{
a dataframe
}
\description{
Convert RAP-ID to MSU-ID
}
\details{
Convert RAP-ID to MSU-ID
}
