% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot}
\usage{
volcano_plot(DEG, logFC = 2, q.value = 0.05)
}
\arguments{
\item{DEG}{needs analyzed list for drawing volcano plot}

\item{logFC}{needs cutoff log2 fold change value, the default value is absolute value 2}

\item{q.value}{needs cutoff q.value, the default value is absolute value 0.05}
}
\value{
a graph
}
\description{
Draw volcano plot
}
\details{
Draw volcano plot
}
